Initial sparse global order reader
This adds a basic reader for the sparse global order case. The reader
uses a merge like algorithm to process tiles instead of sorting through
all the coordinates. It is hidden behind a configuration option but
when enabled, all unit tests are passing.

In this first version, there is very little memory budget management
and serialization is not yet supported.

---
TYPE: IMPROVEMENT
DESC: Initial sparse global order reader
