+++ test/src/unit-capi-incomplete-2.cc
@@ -1,5 +1,5 @@
 /**
- * @file   unit-capi-incomplete.cc
+ * @file   unit-capi-incomplete-2.cc
  *
  * @section LICENSE
  *
