+++ test/src/unit-cppapi-hilbert.cc
@@ -31,6 +31,7 @@
  */
 
 #include "catch.hpp"
+#include "test/src/helpers.h"
 #include "tiledb/sm/cpp_api/tiledb"
 
 using namespace tiledb;
@@ -403,7 +404,10 @@ TEST_CASE(
     CHECK_NOTHROW(query_r.submit());
 
     // Check results again
-    CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
+    CHECK(
+        query_r.query_status() == (test::use_refactored_readers() ?
+                                       Query::Status::COMPLETE :
+                                       Query::Status::INCOMPLETE));
     CHECK(query_r.result_buffer_elements()["a"].second == 2);
     c_buff_a = {4, 1};
     c_buff_d1 = {5, 4};
@@ -412,10 +416,16 @@ TEST_CASE(
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_buff_d2 == c_buff_d2);
 
-    // Read until complete
-    CHECK_NOTHROW(query_r.submit());
-    CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    /**
+     * Old reader needs an extra round here to finish processing all the
+     * partitions in the subarray. New reader is done earlier.
+     */
+    if (!test::use_refactored_readers()) {
+      // Read until complete
+      CHECK_NOTHROW(query_r.submit());
+      CHECK(query_r.query_status() == Query::Status::COMPLETE);
+      CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    }
 
     array_r.close();
   }
@@ -856,7 +866,10 @@ TEST_CASE(
     CHECK_NOTHROW(query_r.submit());
 
     // Check results again
-    CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
+    CHECK(
+        query_r.query_status() == (test::use_refactored_readers() ?
+                                       Query::Status::COMPLETE :
+                                       Query::Status::INCOMPLETE));
     CHECK(query_r.result_buffer_elements()["a"].second == 2);
     c_buff_a = {4, 1};
     c_buff_d1 = {-45, -46};
@@ -865,10 +878,16 @@ TEST_CASE(
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_buff_d2 == c_buff_d2);
 
-    // Read until complete
-    CHECK_NOTHROW(query_r.submit());
-    CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    /**
+     * Old reader needs an extra round here to finish processing all the
+     * partitions in the subarray. New reader is done earlier.
+     */
+    if (!test::use_refactored_readers()) {
+      // Read until complete
+      CHECK_NOTHROW(query_r.submit());
+      CHECK(query_r.query_status() == Query::Status::COMPLETE);
+      CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    }
 
     array_r.close();
   }
@@ -1262,7 +1281,10 @@ TEST_CASE(
     CHECK_NOTHROW(query_r.submit());
 
     // Check results again
-    CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
+    CHECK(
+        query_r.query_status() == (test::use_refactored_readers() ?
+                                       Query::Status::COMPLETE :
+                                       Query::Status::INCOMPLETE));
     CHECK(query_r.result_buffer_elements()["a"].second == 2);
     c_buff_a = {1, 4};
     c_buff_d1 = {0.41f, 0.4f};
@@ -1271,10 +1293,16 @@ TEST_CASE(
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_buff_d2 == c_buff_d2);
 
-    // Read until complete
-    CHECK_NOTHROW(query_r.submit());
-    CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    /**
+     * Old reader needs an extra round here to finish processing all the
+     * partitions in the subarray. New reader is done earlier.
+     */
+    if (!test::use_refactored_readers()) {
+      // Read until complete
+      CHECK_NOTHROW(query_r.submit());
+      CHECK(query_r.query_status() == Query::Status::COMPLETE);
+      CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    }
 
     array_r.close();
   }
@@ -1307,7 +1335,10 @@ TEST_CASE(
     CHECK_NOTHROW(query_r.submit());
 
     // Check results again
-    CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
+    CHECK(
+        query_r.query_status() == (test::use_refactored_readers() ?
+                                       Query::Status::COMPLETE :
+                                       Query::Status::INCOMPLETE));
     CHECK(query_r.result_buffer_elements()["a"].second == 2);
     c_buff_a = {1, 4};
     c_buff_d1 = {0.41f, 0.4f};
@@ -1316,10 +1347,16 @@ TEST_CASE(
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_buff_d2 == c_buff_d2);
 
-    // Read until complete
-    CHECK_NOTHROW(query_r.submit());
-    CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    /**
+     * Old reader needs an extra round here to finish processing all the
+     * partitions in the subarray. New reader is done earlier.
+     */
+    if (!test::use_refactored_readers()) {
+      // Read until complete
+      CHECK_NOTHROW(query_r.submit());
+      CHECK(query_r.query_status() == Query::Status::COMPLETE);
+      CHECK(query_r.result_buffer_elements()["a"].second == 0);
+    }
 
     array_r.close();
   }
@@ -1980,17 +2017,37 @@ TEST_CASE(
     // (33, t1)       ->     877430626372812800
     // (1a, cat)      ->     919167533801450154
     CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 2);
     r_buff_d1.resize(query_r.result_buffer_elements()["d1"].second);
     r_buff_d2.resize(query_r.result_buffer_elements()["d2"].second);
     r_off_d1.resize(query_r.result_buffer_elements()["d1"].first);
     r_off_d2.resize(query_r.result_buffer_elements()["d2"].first);
     r_buff_a.resize(query_r.result_buffer_elements()["a"].second);
-    std::vector<int32_t> c_buff_a = {3, 2};
-    std::string c_buff_d1("dogcamel");
-    std::vector<uint64_t> c_off_d1 = {0, 3};
-    std::string c_buff_d2("stopstock");
-    std::vector<uint64_t> c_off_d2 = {0, 4};
+    std::vector<int32_t> c_buff_a;
+    std::string c_buff_d1;
+    std::vector<uint64_t> c_off_d1;
+    std::string c_buff_d2;
+    std::vector<uint64_t> c_off_d2;
+
+    /**
+     * Refactored reader tries to fill as much as possible.
+     * Old reader splits partition in two.
+     */
+    if (test::use_refactored_readers()) {
+      CHECK(query_r.result_buffer_elements()["a"].second == 3);
+      c_buff_a = {3, 2, 1};
+      c_buff_d1 = std::string("dogcamel33");
+      c_off_d1 = {0, 3, 8};
+      c_buff_d2 = std::string("stopstockt1");
+      c_off_d2 = {0, 4, 9};
+    } else {
+      CHECK(query_r.result_buffer_elements()["a"].second == 2);
+      c_buff_a = {3, 2};
+      c_buff_d1 = std::string("dogcamel");
+      c_off_d1 = {0, 3};
+      c_buff_d2 = std::string("stopstock");
+      c_off_d2 = {0, 4};
+    }
+
     CHECK(r_buff_a == c_buff_a);
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_off_d1 == c_off_d1);
@@ -2006,17 +2063,28 @@ TEST_CASE(
     // (33, t1)       ->     877430626372812800
     // (1a, cat)      ->     919167533801450154
     CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 2);
     r_buff_d1.resize(query_r.result_buffer_elements()["d1"].second);
     r_buff_d2.resize(query_r.result_buffer_elements()["d2"].second);
     r_off_d1.resize(query_r.result_buffer_elements()["d1"].first);
     r_off_d2.resize(query_r.result_buffer_elements()["d2"].first);
     r_buff_a.resize(query_r.result_buffer_elements()["a"].second);
-    c_buff_a = {1, 4};
-    c_buff_d1 = std::string("331a");
-    c_off_d1 = {0, 2};
-    c_buff_d2 = std::string("t1cat");
-    c_off_d2 = {0, 2};
+
+    if (test::use_refactored_readers()) {
+      CHECK(query_r.result_buffer_elements()["a"].second == 1);
+      c_buff_a = {4};
+      c_buff_d1 = std::string("1a");
+      c_off_d1 = {0};
+      c_buff_d2 = std::string("cat");
+      c_off_d2 = {0};
+    } else {
+      CHECK(query_r.result_buffer_elements()["a"].second == 2);
+      c_buff_a = {1, 4};
+      c_buff_d1 = std::string("331a");
+      c_off_d1 = {0, 2};
+      c_buff_d2 = std::string("t1cat");
+      c_off_d2 = {0, 2};
+    }
+
     CHECK(r_buff_a == c_buff_a);
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_off_d1 == c_off_d1);
@@ -2051,17 +2119,37 @@ TEST_CASE(
     // (33, t1)       ->     877430626372812800
     // (1a, cat)      ->     919167533801450154
     CHECK(query_r.query_status() == Query::Status::INCOMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 2);
     r_buff_d1.resize(query_r.result_buffer_elements()["d1"].second);
     r_buff_d2.resize(query_r.result_buffer_elements()["d2"].second);
     r_off_d1.resize(query_r.result_buffer_elements()["d1"].first);
     r_off_d2.resize(query_r.result_buffer_elements()["d2"].first);
     r_buff_a.resize(query_r.result_buffer_elements()["a"].second);
-    std::vector<int32_t> c_buff_a = {3, 2};
-    std::string c_buff_d1("dogcamel");
-    std::vector<uint64_t> c_off_d1 = {0, 3};
-    std::string c_buff_d2("stopstock");
-    std::vector<uint64_t> c_off_d2 = {0, 4};
+    std::vector<int32_t> c_buff_a;
+    std::string c_buff_d1;
+    std::vector<uint64_t> c_off_d1;
+    std::string c_buff_d2;
+    std::vector<uint64_t> c_off_d2;
+
+    /**
+     * Refactored reader tries to fill as much as possible.
+     * Old reader splits partition in two.
+     */
+    if (test::use_refactored_readers()) {
+      CHECK(query_r.result_buffer_elements()["a"].second == 3);
+      c_buff_a = {3, 2, 1};
+      c_buff_d1 = std::string("dogcamel33");
+      c_off_d1 = {0, 3, 8};
+      c_buff_d2 = std::string("stopstockt1");
+      c_off_d2 = {0, 4, 9};
+    } else {
+      CHECK(query_r.result_buffer_elements()["a"].second == 2);
+      c_buff_a = {3, 2};
+      c_buff_d1 = std::string("dogcamel");
+      c_off_d1 = {0, 3};
+      c_buff_d2 = std::string("stopstock");
+      c_off_d2 = {0, 4};
+    }
+
     CHECK(r_buff_a == c_buff_a);
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_off_d1 == c_off_d1);
@@ -2077,17 +2165,28 @@ TEST_CASE(
     // (33, t1)       ->     877430626372812800
     // (1a, cat)      ->     919167533801450154
     CHECK(query_r.query_status() == Query::Status::COMPLETE);
-    CHECK(query_r.result_buffer_elements()["a"].second == 2);
     r_buff_d1.resize(query_r.result_buffer_elements()["d1"].second);
     r_buff_d2.resize(query_r.result_buffer_elements()["d2"].second);
     r_off_d1.resize(query_r.result_buffer_elements()["d1"].first);
     r_off_d2.resize(query_r.result_buffer_elements()["d2"].first);
     r_buff_a.resize(query_r.result_buffer_elements()["a"].second);
-    c_buff_a = {1, 4};
-    c_buff_d1 = std::string("331a");
-    c_off_d1 = {0, 2};
-    c_buff_d2 = std::string("t1cat");
-    c_off_d2 = {0, 2};
+
+    if (test::use_refactored_readers()) {
+      CHECK(query_r.result_buffer_elements()["a"].second == 1);
+      c_buff_a = {4};
+      c_buff_d1 = std::string("1a");
+      c_off_d1 = {0};
+      c_buff_d2 = std::string("cat");
+      c_off_d2 = {0};
+    } else {
+      CHECK(query_r.result_buffer_elements()["a"].second == 2);
+      c_buff_a = {1, 4};
+      c_buff_d1 = std::string("331a");
+      c_off_d1 = {0, 2};
+      c_buff_d2 = std::string("t1cat");
+      c_off_d2 = {0, 2};
+    }
+
     CHECK(r_buff_a == c_buff_a);
     CHECK(r_buff_d1 == c_buff_d1);
     CHECK(r_off_d1 == c_off_d1);
