+++ tiledb/sm/c_api/tiledb.h
@@ -1044,6 +1044,9 @@ TILEDB_EXPORT void tiledb_config_free(tiledb_config_t** config);
  * - `sm.use_refactored_readers` <br>
  *    Use the refactored readers or not. <br>
  *    **Default**: false
+ * - `sm.mem.total_budget` <br>
+ *    Memory budget for the sparse global order reader. <br>
+ *    **Default**: \10GB
  * - `vfs.read_ahead_size` <br>
  *    The maximum byte size to read-ahead from the backend. <br>
  *    **Default**: 102400
