+++ tiledb/sm/config/config.h
@@ -157,6 +157,9 @@ class Config {
   /** Whether or not to use the refactored readers. */
   static const std::string SM_USE_REFACTORED_READERS;
 
+  /** Maximum memory budget for the sparse global order reader. */
+  static const std::string SM_MEM_TOTAL_BUDGET;
+
   /** Whether or not the signal handlers are installed. */
   static const std::string SM_ENABLE_SIGNAL_HANDLERS;
