+++ tiledb/sm/cpp_api/config.h
@@ -392,6 +392,9 @@ class Config {
    * - `sm.use_refactored_readers` <br>
    *    Use the refactored readers or not. <br>
    *    **Default**: false
+   * - `sm.mem.total_budget` <br>
+   *    Memory budget for the sparse global order reader. <br>
+   *    **Default**: \10GB
    * - `vfs.read_ahead_size` <br>
    *    The maximum byte size to read-ahead from the backend. <br>
    *    **Default**: 102400
